+++ client/components/Postman/Postman.js
@@ -35,35 +35,39 @@ const {
   checkNameIsExistInArray
 } = require('common/postmanLib.js');
 
+const plugin = require('client/plugin.js');
+
+const createContext = require('common/createContext')
+
 const HTTP_METHOD = constants.HTTP_METHOD;
 const InputGroup = Input.Group;
 const Option = Select.Option;
 const Panel = Collapse.Panel;
 
-const InsertCodeMap = [
+export const InsertCodeMap = [
   {
     code: 'assert.equal(status, 200)',
     title: '断言 httpCode 等于 200'
   },
   {
-    code: 'assert.equal(body.errcode, 0)',
-    title: '断言返回数据 errcode 是 0'
+    code: 'assert.equal(body.code, 0)',
+    title: '断言返回数据 code 是 0'
   },
   {
     code: 'assert.notEqual(status, 404)',
     title: '断言 httpCode 不是 404'
   },
   {
-    code: 'assert.notEqual(body.errcode, 40000)',
-    title: '断言返回数据 errcode 不是 40000'
+    code: 'assert.notEqual(body.code, 40000)',
+    title: '断言返回数据 code 不是 40000'
   },
   {
-    code: 'assert.deepEqual(body, {"errcode": 0})',
-    title: '断言对象 body 等于 {"errcode": 0}'
+    code: 'assert.deepEqual(body, {"code": 0})',
+    title: '断言对象 body 等于 {"code": 0}'
   },
   {
-    code: 'assert.notDeepEqual(body, {"errcode": 0})',
-    title: '断言对象 body 不等于 {"errcode": 0}'
+    code: 'assert.notDeepEqual(body, {"code": 0})',
+    title: '断言对象 body 不等于 {"code": 0}'
   }
 ];
 
@@ -104,12 +108,14 @@ ParamsNameComponent.propTypes = {
   desc: PropTypes.string,
   name: PropTypes.string
 };
-
 export default class Run extends Component {
   static propTypes = {
     data: PropTypes.object, //接口原有数据
     save: PropTypes.func, //保存回调方法
-    type: PropTypes.string //enum[case, inter], 判断是在接口页面使用还是在测试集
+    type: PropTypes.string, //enum[case, inter], 判断是在接口页面使用还是在测试集
+    curUid: PropTypes.number.isRequired,
+    interfaceId: PropTypes.number.isRequired,
+    projectId: PropTypes.number.isRequired
   };
 
   constructor(props) {
@@ -129,8 +135,16 @@ export default class Run extends Component {
       envModalVisible: false,
       test_res_header: null,
       test_res_body: null,
+      autoPreviewHTML: true,
       ...this.props.data
     };
+  }
+
+  get testResponseBodyIsHTML() {
+    const hd = this.state.test_res_header
+    return hd != null
+      && typeof hd === 'object'
+      && String(hd['Content-Type'] || hd['content-type']).indexOf('text/html') !== -1
   }
 
   checkInterfaceData(data) {
@@ -202,12 +216,33 @@ export default class Run extends Component {
       body = JSON.stringify(result.data);
     }
 
+    let example = {}
+    if(this.props.type === 'inter'){
+      example = ['req_headers', 'req_query', 'req_body_form'].reduce(
+        (res, key) => {
+          res[key] = (data[key] || []).map(item => {
+            if (
+              item.type !== 'file' // 不是文件类型
+                && (item.value == null || item.value === '') // 初始值为空
+                && item.example != null // 有示例值
+            ) {
+              item.value = item.example;
+            }
+            return item;
+          })
+          return res;
+        },
+        {}
+      )
+    }
+
     this.setState(
       {
         ...this.state,
         test_res_header: null,
         test_res_body: null,
         ...data,
+        ...example,
         req_body_other: body,
         resStatusCode: null,
         test_valid_msg: null,
@@ -299,8 +334,29 @@ export default class Run extends Component {
     let options = handleParams(this.state, this.handleValue),
       result;
 
+
+    await plugin.emitHook('before_request', options, {
+      type: this.props.type,
+      caseId: options.caseId,
+      projectId: this.props.projectId,
+      interfaceId: this.props.interfaceId
+    });
+
     try {
-      result = await crossRequest(options, this.state.pre_script, this.state.after_script);
+      options.taskId = this.props.curUid;
+      result = await crossRequest(options, options.pre_script || this.state.pre_script, options.after_script || this.state.after_script, createContext(
+        this.props.curUid,
+        this.props.projectId,
+        this.props.interfaceId
+      ));
+
+      await plugin.emitHook('after_request', result, {
+        type: this.props.type,
+        caseId: options.caseId,
+        projectId: this.props.projectId,
+        interfaceId: this.props.interfaceId
+      });
+
       result = {
         header: result.res.header,
         body: result.res.body,
@@ -308,6 +364,7 @@ export default class Run extends Component {
         statusText: result.res.statusText,
         runTime: result.runTime
       };
+
     } catch (data) {
       result = {
         header: data.header,
@@ -815,13 +872,14 @@ export default class Run extends Component {
                         )}
                         <span className="eq-symbol">=</span>
                         {item.type === 'file' ? (
-                          <Input
-                            type="file"
-                            id={'file_' + index}
-                            onChange={e => this.changeBody(e.target.value, index, 'value')}
-                            multiple
-                            className="value"
-                          />
+                          '因Chrome最新版安全策略限制，不再支持文件上传'
+                          // <Input
+                          //   type="file"
+                          //   id={'file_' + index}
+                          //   onChange={e => this.changeBody(e.target.value, index, 'value')}
+                          //   multiple
+                          //   className="value"
+                          // />
                         ) : (
                           <Input
                             value={item.value}
@@ -875,6 +933,9 @@ export default class Run extends Component {
               >
                 {this.state.resStatusCode + '  ' + this.state.resStatusText}
               </h2>
+              <div>
+                <a rel="noopener noreferrer"  target="_blank" href="https://juejin.im/post/5c888a3e5188257dee0322af">YApi 新版如何查看 http 请求数据</a>
+              </div>
               {this.state.test_valid_msg && (
                 <Alert
                   message={
@@ -914,14 +975,25 @@ export default class Run extends Component {
                 <div className="body">
                   <div className="container-title">
                     <h4>Body</h4>
+                    <Checkbox
+                      checked={this.state.autoPreviewHTML}
+                      onChange={e => this.setState({ autoPreviewHTML: e.target.checked })}>
+                      <span>自动预览HTML</span>
+                    </Checkbox>
                   </div>
-                  <AceEditor
-                    readOnly={true}
-                    className="pretty-editor-body"
-                    data={this.state.test_res_body}
-                    mode={handleContentType(this.state.test_res_header)}
-                    // mode="html"
-                  />
+                  {
+                    this.state.autoPreviewHTML && this.testResponseBodyIsHTML
+                      ? <iframe
+                          className="pretty-editor-body"
+                          srcDoc={this.state.test_res_body}
+                        />
+                      : <AceEditor
+                          readOnly={true}
+                          className="pretty-editor-body"
+                          data={this.state.test_res_body}
+                          mode={handleContentType(this.state.test_res_header)}
+                      />
+                  }
                 </div>
               </div>
             </Spin>