+++ server/utils/commons.js
@@ -14,16 +14,18 @@ const _ = require('underscore');
 const Ajv = require('ajv');
 const Mock = require('mockjs');
 
+
+
 const ejs = require('easy-json-schema');
 
 const jsf = require('json-schema-faker');
-const formats = require('../../common/formats');
+const { schemaValidator } = require('../../common/utils');
 const http = require('http');
 
-jsf.extend ('mock', function () {
+jsf.extend('mock', function () {
   return {
     mock: function (xx) {
-      return Mock.mock (xx);
+      return Mock.mock(xx);
     }
   };
 });
@@ -43,9 +45,9 @@ const defaultOptions = {
 //   });
 // });
 
-exports.schemaToJson = function(schema, options = {}) {
+exports.schemaToJson = function (schema, options = {}) {
   Object.assign(options, defaultOptions);
-  
+
   jsf.option(options);
   let result;
   try {
@@ -181,7 +183,7 @@ exports.sendMail = (options, cb) => {
 
   cb =
     cb ||
-    function(err) {
+    function (err) {
       if (err) {
         yapi.commons.log('send mail ' + options.to + ' error,' + err.message, 'error');
       } else {
@@ -251,7 +253,7 @@ exports.handleVarPath = (pathname, params) => {
       }
     }
   }
-  pathname.replace(/\{(.+?)\}/g, function(str, match) {
+  pathname.replace(/\{(.+?)\}/g, function (str, match) {
     insertParams(match);
   });
 };
@@ -279,15 +281,18 @@ exports.verifyPath = path => {
  * a = {a: 2}
  */
 exports.sandbox = (sandbox, script) => {
-  const vm = require('vm');
-  sandbox = sandbox || {};
-  script = new vm.Script(script);
-  const context = new vm.createContext(sandbox);
-  script.runInContext(context, {
-    timeout: 3000
-  });
-
-  return sandbox;
+  try {
+    const vm = require('vm');
+    sandbox = sandbox || {};	
+    script = new vm.Script(script);	
+    const context = new vm.createContext(sandbox);	
+    script.runInContext(context, {	
+      timeout: 3000	
+    });	      
+    return sandbox
+  } catch (err) {
+    throw err
+  }
 };
 
 function trim(str) {
@@ -417,8 +422,9 @@ exports.createAction = (router, baseurl, routerController, action, path, method,
     let inst = new routerController(ctx);
     try {
       await inst.init(ctx);
+      ctx.params = Object.assign({}, ctx.request.query, ctx.request.body, ctx.params);
       if (inst.schemaMap && typeof inst.schemaMap === 'object' && inst.schemaMap[action]) {
-        ctx.params = Object.assign({}, ctx.request.query, ctx.request.body, ctx.params);
+
         let validResult = yapi.commons.validateParams(inst.schemaMap[action], ctx.params);
 
         if (!validResult.valid) {
@@ -450,7 +456,7 @@ function handleParamsValue(params, val) {
   let value = {};
   try {
     params = params.toObject();
-  } catch (e) {}
+  } catch (e) { }
   if (params.length === 0 || val.length === 0) {
     return params;
   }
@@ -509,18 +515,19 @@ function convertString(variable) {
     return variable.name + ': ' + variable.message;
   }
   try {
+    if (variable && typeof variable === 'string') {
+      return variable;
+    }
     return JSON.stringify(variable, null, '   ');
   } catch (err) {
     return variable || '';
   }
 }
 
-exports.runCaseScript = async function runCaseScript(params) {
-  let script = params.script;
-  // script 是断言
-  if (!script) {
-    return yapi.commons.resReturn('ok');
-  }
+
+exports.runCaseScript = async function runCaseScript(params, colId, interfaceId) {
+  const colInst = yapi.getInst(interfaceColModel);
+  let colData = await colInst.get(colId);
   const logs = [];
   const context = {
     assert: require('assert'),
@@ -536,14 +543,56 @@ exports.runCaseScript = async function runCaseScript(params) {
 
   let result = {};
   try {
-    result = yapi.commons.sandbox(context, script);
 
+    if (colData.checkHttpCodeIs200) {
+      let status = +params.response.status;
+      if (status !== 200) {
+        throw ('Http status code 不是 200，请检查(该规则来源于于 [测试集->通用规则配置] )')
+      }
+    }
+
+    if (colData.checkResponseField.enable) {
+      if (params.response.body[colData.checkResponseField.name] != colData.checkResponseField.value) {
+        throw (`返回json ${colData.checkResponseField.name} 值不是${colData.checkResponseField.value}，请检查(该规则来源于于 [测试集->通用规则配置] )`)
+      }
+    }
+
+    if (colData.checkResponseSchema) {
+      const interfaceInst = yapi.getInst(interfaceModel);
+      let interfaceData = await interfaceInst.get(interfaceId);
+      if (interfaceData.res_body_is_json_schema && interfaceData.res_body) {
+        let schema = JSON.parse(interfaceData.res_body);
+        let result = schemaValidator(schema, context.body)
+        if (!result.valid) {
+          throw (`返回Json 不符合 response 定义的数据结构,原因: ${result.message}
+数据结构如下：
+${JSON.stringify(schema, null, 2)}`)
+        }
+      }
+    }
+
+    if (colData.checkScript.enable) {
+      let globalScript = colData.checkScript.content;
+      // script 是断言
+      if (globalScript) {
+        logs.push('执行脚本：' + globalScript)
+        result = yapi.commons.sandbox(context, globalScript);
+      }
+    }
+
+
+    let script = params.script;
+    // script 是断言
+    if (script) {
+      logs.push('执行脚本:' + script)
+      result = yapi.commons.sandbox(context, script);
+    }
     result.logs = logs;
     return yapi.commons.resReturn(result);
   } catch (err) {
     logs.push(convertString(err));
     result.logs = logs;
-
+    logs.push(err.name + ': ' + err.message)
     return yapi.commons.resReturn(result, 400, err.name + ': ' + err.message);
   }
 };
@@ -562,44 +611,9 @@ exports.getUserdata = async function getUserdata(uid, role) {
     email: userData.email
   };
 };
-// 邮件发送
-exports.sendNotice = async function(projectId, data) {
-  const followInst = yapi.getInst(followModel);
-  const userInst = yapi.getInst(userModel);
-  const projectInst = yapi.getInst(projectModel);
-  const list = await followInst.listByProjectId(projectId);
-  const starUsers = list.map(item => item.uid);
-
-  const projectList = await projectInst.get(projectId);
-  const projectMenbers = projectList.members
-    .filter(item => item.email_notice)
-    .map(item => item.uid);
-
-  const users = arrUnique(projectMenbers, starUsers);
-  const usersInfo = await userInst.findByUids(users);
-  const emails = usersInfo.map(item => item.email).join(',');
-
-  try {
-    yapi.commons.sendMail({
-      to: emails,
-      contents: data.content,
-      subject: data.title
-    });
-  } catch (e) {
-    yapi.commons.log('邮件发送失败：' + e, 'error');
-  }
-};
-
-function arrUnique(arr1, arr2) {
-  let arr = arr1.concat(arr2);
-  let res = arr.filter(function(item, index, arr) {
-    return arr.indexOf(item) === index;
-  });
-  return res;
-}
 
 // 处理mockJs脚本
-exports.handleMockScript = function(script, context) {
+exports.handleMockScript = function (script, context) {
   let sandbox = {
     header: context.ctx.header,
     query: context.ctx.query,
@@ -614,7 +628,7 @@ exports.handleMockScript = function(script, context) {
   sandbox.cookie = {};
 
   context.ctx.header.cookie &&
-    context.ctx.header.cookie.split(';').forEach(function(Cookie) {
+    context.ctx.header.cookie.split(';').forEach(function (Cookie) {
       var parts = Cookie.split('=');
       sandbox.cookie[parts[0].trim()] = (parts[1] || '').trim();
     });
@@ -629,8 +643,8 @@ exports.handleMockScript = function(script, context) {
 
 
 
-exports.createWebAPIRequest = function(ops) {
-  return new Promise(function(resolve, reject) {
+exports.createWebAPIRequest = function (ops) {
+  return new Promise(function (resolve, reject) {
     let req = '';
     let http_client = http.request(
       {
@@ -639,25 +653,25 @@ exports.createWebAPIRequest = function(ops) {
         port: ops.port,
         path: ops.path
       },
-      function(res) {
-        res.on('error', function(err) {
+      function (res) {
+        res.on('error', function (err) {
           reject(err);
         });
         res.setEncoding('utf8');
         if (res.statusCode != 200) {
-          reject({message: 'statusCode != 200'});
+          reject({ message: 'statusCode != 200' });
         } else {
-          res.on('data', function(chunk) {
+          res.on('data', function (chunk) {
             req += chunk;
           });
-          res.on('end', function() {
+          res.on('end', function () {
             resolve(req);
           });
         }
       }
     );
     http_client.on('error', (e) => {
-      reject({message: `request error: ${e.message}`});
+      reject({ message: `request error: ${e.message}` });
     });
     http_client.end();
   });