+++ static/prd/assets.js
@@ -1,1 +1,5 @@
+<<<<<<< HEAD
 window.WEBPACK_ASSETS = {"manifest":{"js":"manifest@f2f4bd774d6c221b3d5f.js"},"index.js":{"js":"index@e0dcaf0f9d277ddb5a86.js","css":"index@e0dcaf0f9d277ddb5a86.css"},"lib":{"js":"lib@771cb7fb7ef62c82c3d7.js"},"lib2":{"js":"lib2@2faaf36c45579941942d.js"},"lib3":{"js":"lib3@bb3a29cee0a71789f058.js"},"":{"gz":["lib@771cb7fb7ef62c82c3d7.js.gz","lib2@2faaf36c45579941942d.js.gz","index@e0dcaf0f9d277ddb5a86.css.gz","lib3@bb3a29cee0a71789f058.js.gz","index@e0dcaf0f9d277ddb5a86.js.gz"]}}
+=======
+window.WEBPACK_ASSETS = {"index.js":{"js":"index@4df3f8796e0eea779ba6.js","css":"index@4df3f8796e0eea779ba6.css"},"lib":{"js":"lib@aaed0f9bfcc4ff921c5a.js"},"lib2":{"js":"lib2@819909a89ca45b94d9bf.js"},"lib3":{"js":"lib3@93b22d7460620277a5da.js"},"manifest":{"js":"manifest@f2f4bd774d6c221b3d5f.js"}}
+>>>>>>> master