+++ CHANGELOG.md
@@ -1,3 +1,127 @@
+## v1.9.2
+* fix: 修复高级 mock 无效的bug
+* opti: 对登录 email 空格的过滤
+* fix: 修复deepMath 增加对原型属性比较
+
+## v1.9.1
+* 修复因 mongodb 一个废弃报错导致部署不成功问题
+
+## v1.9.0
+* 修复测试集合部分情况下闪动问题
+* 修改ldap filter的匹配规则，使其可以匹配&和|操作符  Merge pull request #1631 from vvkkhjt/master
+* support switch_notice for /interface/save  Merge pull request #1646 from tangcent/feature/openapi_notice
+* 自动化测试时服务端测试node默认2分钟没有返回就直接断掉连接，可以手动设置一个超时时间 Merge pull request #1675 from liugddx/master
+* 修复导出的swagger.json 中 required 一直是 false 的 bug
+* 修复 schema2json传required参数时，导致faker失败
+* 修改ldap filter的匹配规则 
+* 更新 sm2tsservice 3.2.0及以后版本的使用配置
+
+## v1.8.8
+* 更新了 cross-request [教程](https://juejin.im/post/5e3bbd986fb9a07ce152b53d)，发布了最新的 3.1 版本 
+
+## v1.8.7
+* 因 chrome 官方下架了 yapi 扩展，整理了本地安装教程
+
+## v1.8.6
+* 优化 swagger 文档导入分类策略，优先使用根路径的 tags 做分类，避免出现特别多分类的问题
+
+## v1.8.5
+* 改善 swagger 自动导入，不再支持秒级别的 cron 表达式,默认使用 10分钟更新一次的频率
+* 修复输入空的 swagger地址 ，会发起请求的 bug
+* 优化 swagger 数据导入，不会导入空的分类，不会使用版本号作为分类名称
+* swagger 导入自动增加 basePath 到项目配置
+
+## v1.8.4
+* 修复 swagger 导入数据时，如果数据格式中缺少 in 的参数，会丢失请求参数
+* 修复当传入数据格式的 method 不规范时，容易导致获取对象为空，出现异常导致动态页面无法打开
+* 修复解决 json-schema-faker/json-schema-faker#453 问题
+
+## v1.8.3
+* 修复管理员无法看到所有分组的 bug
+
+## v1.8.2
+* 重构分组列表功能实现，大幅度优化首屏加载速度
+* 接口运行界面设置 header、query、form 的初始值为其示例值
+* 运行接口请求时支持自动预览HTML
+
+### v1.8.1
+* 优化插件【Swagger 自动同步】在添加地址时的服务端校验行为
+* 优化单个测试用例执行超时时间限制,从3秒改为10秒
+
+### v1.8.0
+* filtering interface on the server instead of client
+
+### v1.7.2
+* 支持接口路径模糊搜索，不包含 basepath
+
+### v1.7.1
+* 废弃 yapi.ymfe.org 文档站点
+
+### v1.7.0
+* fix：修复md两个undefined以及run_auto_test中执行用例id问题 #1024
+
+### v1.7.0-beta.1
+* 修复storage保存逻辑错误
+
+### v1.7.0-beta.0
+* **[插件]** 新增默认插件，支持通过 token 导出包含 basepath 的 json 格式接口，并整合添加 sm2tsservice 入口
+* **[插件]** 新增默认插件，支持swagger数据同步
+* 修复不兼容 node7.6 bug
+
+### v1.5.14 
+* 修复接口运行部分请求参数默认使用示例填写值导致无法删除参数bug
+* 修复无法保存 global bug
+
+### v1.5.13 （存在bug）
+* 支持 pre-script 脚本持久化数据存储，storage 兼容浏览器和服务端，并且是持久化数据存储，不会丢失，用法类似于 localStorage
+* 修复了swagger 数据导入bug
+* 修复接口运行部分请求参数默认使用示例填写值导致无法删除参数bug
+
+### v1.5.12  （存在bug）
+* 废弃 v1.6.x 新增功能，因为有不可控的bug出现
+* 支持项目设置 hook
+* 开放api 新增 '/api/plugin/export'
+* 接口运行部分请求参数默认使用示例填写值
+
+### v1.5.10
+* 解决 license should be a valid SPDX license expression 报错
+* 修改OpenAPI比较版本方法
+* fix复制路径不包含基本路径
+* 修复了第一次部署，首页一直处于 loading bug
+
+### v1.5.7
+* 数据导入默认使用完全覆盖
+* 升级新版本 cross-request 扩展，因 chrome 安全策略限制，不再支持文件上传
+* fix 重复的 moment 依赖，导致安装时报错
+* feat: add jsrsasign Lib
+
+### v1.5.6
+* 修复 /api/open/import_data 参数bug
+* 修复  /api/open/import_data 文档错误，merge 参数误写为 dataSync
+
+### v1.5.5
+* cross-request 升级到 2.10
+* /api/open/import_data 新增 url 参数，支持服务端 url 导入
+
+### v1.5.2
+* 新增 openapi `/api/project/get`，可获取项目基本信息
+
+### v1.5.1
+
+* 优化 restful api 动态路由权重匹配算法，匹配更加精确
+* openapi 新增 `/api/interface/list_cat`，获取某个分类下所有接口
+* 新增了 rap数据导入到 yapi 插件 [rap2yapi](https://github.com/wxxcarl/yapi-plugin-import-rap)
+
+### v1.5.0
+
+* 优化开放 api功能，现在 token 带有用户信息了
+* 修复无法获取请求302 跳转前的 headers
+
+### v1.4.4
+* 优化了 json-schema 编辑器交互，修复了参数名写到一半提示重复的问题
+* 优化了首页体验，提升页面打开速度
+* 新增自动化测试通用规则配置功能
+
 ### v1.4.3
 * 修复了可视化安装，mongodb 报错的问题
 * 支持了 swagger 导出功能
@@ -499,3 +623,7 @@
 
 * 接口备注集成了富文本编辑
 * 支持 har 协议的接口数据导入
+
+
+todo:
+新增 crypto 加密函数